+++ src/services/shipment_service.js
@@ -68,7 +68,7 @@ export default (easypostClient) =>
       try {
         const response = await easypostClient._post(url, wrappedParams);
 
-        return this._convertToEasyPostObject(response.body);
+        return this._convertToEasyPostObject(response.body, wrappedParams);
       } catch (e) {
         return Promise.reject(e);
       }
@@ -88,7 +88,7 @@ export default (easypostClient) =>
       try {
         const response = await easypostClient._get(url, wrappedParams);
 
-        return this._convertToEasyPostObject(response.body);
+        return this._convertToEasyPostObject(response.body, wrappedParams);
       } catch (e) {
         return Promise.reject(e);
       }
@@ -108,7 +108,7 @@ export default (easypostClient) =>
       try {
         const response = await easypostClient._post(url, wrappedParams);
 
-        return this._convertToEasyPostObject(response.body);
+        return this._convertToEasyPostObject(response.body, wrappedParams);
       } catch (e) {
         return Promise.reject(e);
       }
@@ -146,7 +146,7 @@ export default (easypostClient) =>
       try {
         const response = await easypostClient._post(url, wrappedParams);
 
-        return this._convertToEasyPostObject(response.body);
+        return this._convertToEasyPostObject(response.body, wrappedParams);
       } catch (e) {
         return Promise.reject(e);
       }
@@ -172,7 +172,7 @@ export default (easypostClient) =>
       try {
         const response = await easypostClient._post(url, wrappedParams);
 
-        return this._convertToEasyPostObject(response.body);
+        return this._convertToEasyPostObject(response.body, wrappedParams);
       } catch (e) {
         return Promise.reject(e);
       }
@@ -221,11 +221,7 @@ export default (easypostClient) =>
     static async all(params = {}) {
       const url = 'shipments';
 
-      const response = await this._all(url, params);
-      response.purchased = params.purchased;
-      response.include_children = params.include_children;
-
-      return response;
+      return this._all(url, params);
     }
 
     /**
@@ -236,12 +232,8 @@ export default (easypostClient) =>
      */
     static async getNextPage(shipments, pageSize = null) {
       const url = 'shipments';
-      const params = {
-        purchased: shipments.purchased ?? true,
-        include_children: shipments.include_children ?? false,
-      };
 
-      return this._getNextPage(url, shipments, pageSize, params);
+      return this._getNextPage(url, 'shipments', shipments, pageSize);
     }
 
     /**
@@ -272,7 +264,7 @@ export default (easypostClient) =>
       try {
         const response = await easypostClient._get(url, params);
 
-        return this._convertToEasyPostObject(response.body.rates ?? []);
+        return this._convertToEasyPostObject(response.body.rates ?? [], params);
       } catch (e) {
         return Promise.reject(e);
       }